Add hook for mantaining and updating search results
The hook creates a 'Search' function, that marks the input as a possible next query. The actual search query is sent to the backend only after a specified time or if the query string changed too much compared with the last query. The throttler also limits the number of unresolved requests.